+++ path.c
@@ -14,103 +14,6 @@ bool	is_directory(char *path)
 	return (false);
 }
 
-/* 絶対パスを構築する
- * "{g_cwd}/{relative_path}" を返す
- * ex:
- *   - relative_path="dir/dir2/symlink2dir"
- *   - relative_path="./../dir/../././/.//////./"
- */
-char *get_abs_path_from_cwd(char *relative_path)
-{
-	char	**dirs;  // ("..", ".", その他) みたいなのが入ってくる
-	char	*result;
-	char	*tmp;
-	int		i;
-
-	result = ft_strdup(g_cwd);
-	dirs = ft_split(relative_path, '/');
-	if (!dirs)
-		return (NULL);
-	i = 0;
-	while (dirs[i])
-	{
-		tmp = result;
-		result = path_join(tmp, dirs[i]);
-		free(tmp);
-		i++;
-	}
-	free_ptrarr((void **)dirs);
-	// tmp = result;
-	// result = canonicalize_path(tmp);
-	// free(tmp);
-	return (result);
-}
-
-/* パスを正規化する.
- * ex:
- *   - "////hoge//fuga//../././//gaga" -> "/hoge/gaga"
- */
-char	*canonicalize_path(char *path)
-{
-	char	**dirs;
-	char	*result;
-	char	*tmp;
-	int		i;
-
-	dirs = ft_split(path, '/');
-	i = 0;
-	result = ft_strdup("/");
-	while (dirs[i])
-	{
-		if (ft_strncmp(dirs[i], "..", 3) == 0)
-		{
-			tmp = result;
-			result = change_to_parent_dir(result);
-			free(tmp);
-		}
-		else if (ft_strncmp(dirs[i], ".", 2) == 0)
-			;
-		else
-		{
-			tmp = result;
-			result = path_join(tmp, dirs[i]);
-			free(tmp);
-		}
-		PRINT_DEBUG("canon_str: |%s|, dirs[%d]: |%s|\n", result, i, dirs[i]);
-		if (!is_directory(result))
-		{
-			free(result);
-			free_ptrarr((void **)dirs);
-			return (NULL);
-		}
-		i++;
-	}
-	free_ptrarr((void **)dirs);
-	if (ft_strlen(path) >= 2 && path[0] == '/' && path[1] == '/' && path[2] != '/')
-	{
-		tmp = result;
-		result = ft_strjoin("/", result);
-		free(tmp);
-	}
-	return (result);
-}
-
-/* 親ディレクトリに移動した後のパスを返す
- */
-char	*change_to_parent_dir(char *path)
-{
-	char	*result;
-
-	result = ft_substr(path, 0, ft_strrchr(path, '/') - path);
-	// ルートディレクトリ "/" で上の処理を実行すると空文字列が返ってくるので対策
-	if (result && ft_strlen(result) == 0)
-	{
-		free(result);
-		result = ft_strdup("/");
-	}
-	return (result);
-}
-
 /*
  * concatenate dirpath, "/" and filename.
  */
